+++ src/main/scala/mada/peg/Best.scala
@@ -18,8 +18,8 @@ object Shortest {
 }
 
 
-class BestPeg[A](ps: Iterable[Peg[A]], which: (Long, Long) => Long) extends Peg[A] {
-    override def parse(v: Vector[A], first: Long, last: Long) = {
+class BestPeg[A](ps: Iterable[Peg[A]], which: (Int, Int) => Int) extends Peg[A] {
+    override def parse(v: Vector[A], first: Int, last: Int) = {
         var cur = Peg.FAILURE
         for (p <- ps.projection) {
             val i = p.parse(v, first, last)
