+++ src/main/scala/mada/peg/PegProxy.scala
@@ -10,8 +10,8 @@ package mada.peg
 trait PegProxy[A] extends Peg[A] with Proxy {
     override def self: Peg[A]
 
-    override def parse(v: Vector[A], first: Long, last: Long) = self.parse(v, first, last)
-    override def length: Long = self.length
+    override def parse(v: Vector[A], first: Int, last: Int) = self.parse(v, first, last)
+    override def length: Int = self.length
 
 /*
     override def lookBehind = self.lookBehind
