+++ src/main/scala/mada/vec/Cell.scala
@@ -13,7 +13,7 @@ object CellVector {
 
 class CellVector[A](val cell: Cell[A]) extends Vector[A] {
     override def size = 1
-    override def apply(i: Long) = cell.elem
+    override def apply(i: Int) = cell.elem
 
     override def toCell = cell // from-to fusion
 }
