+++ src/main/scala/mada/vec/FileVector.scala
@@ -17,9 +17,9 @@ class CharFileVector private (val randomAccessFile: RandomAccessFile) extends Ve
     def this(name: String, mode: String) = this(new RandomAccessFile(name, mode))
     def close = randomAccessFile.close
 
-    override def size = randomAccessFile.length / 2
-    override def apply(i: Long) = { randomAccessFile.seek(i * 2); randomAccessFile.readChar }
-    override def update(i: Long, e: Char) = { randomAccessFile.seek(i * 2); randomAccessFile.writeChar(e) }
+    override def size = randomAccessFile.length.toInt / 2
+    override def apply(i: Int) = { randomAccessFile.seek(i * 2); randomAccessFile.readChar }
+    override def update(i: Int, e: Char) = { randomAccessFile.seek(i * 2); randomAccessFile.writeChar(e) }
 }
 
 
@@ -30,9 +30,9 @@ class IntFileVector private (val randomAccessFile: RandomAccessFile) extends Vec
     def this(name: String, mode: String) = this(new RandomAccessFile(name, mode))
     def close = randomAccessFile.close
 
-    override def size = randomAccessFile.length / 4
-    override def apply(i: Long) = { randomAccessFile.seek(i * 4); randomAccessFile.readInt }
-    override def update(i: Long, e: Int) = { randomAccessFile.seek(i * 4); randomAccessFile.writeInt(e) }
+    override def size = randomAccessFile.length.toInt / 4
+    override def apply(i: Int) = { randomAccessFile.seek(i * 4); randomAccessFile.readInt }
+    override def update(i: Int, e: Int) = { randomAccessFile.seek(i * 4); randomAccessFile.writeInt(e) }
 }
 
 
@@ -43,7 +43,7 @@ class LongFileVector private (val randomAccessFile: RandomAccessFile) extends Ve
     def this(name: String, mode: String) = this(new RandomAccessFile(name, mode))
     def close = randomAccessFile.close
 
-    override def size = randomAccessFile.length / 8
-    override def apply(i: Long) = { randomAccessFile.seek(i * 8); randomAccessFile.readLong }
-    override def update(i: Long, e: Long) = { randomAccessFile.seek(i * 8); randomAccessFile.writeLong(e) }
+    override def size = randomAccessFile.length.toInt / 8
+    override def apply(i: Int) = { randomAccessFile.seek(i * 8); randomAccessFile.readLong }
+    override def update(i: Int, e: Long) = { randomAccessFile.seek(i * 8); randomAccessFile.writeLong(e) }
 }
